Update 'isRegulation' on initial page load
When the page is loaded for the first time, we only have 'besluitTypeId' as a part of the page url. There is no 'besluitType' object set yet because this object is only set when the current selection in the 'Type Document' menu is manually changed. This is also the case when the page is refreshed after a certain document type has been selected. The issue in this situations would be that we would have 'Reglementen en verordeningen' selected in the 'Type Document' menu but the 'Type reglementen of verordening' menu would not be rendered.