This check is not necessary
The if-statement is most important in the form where they have to fill in the regulationType. There we enforce they can only set a value for regulationType if besluitType == 'Reglement en verorderingen' . When displaying the inzendingen in the overview we shouldn't apply the if-statement anymore, because row.regulationType.label will only contain a value if besluitType == 'Reglement en verorderingen' because the input form forces it like that.