Fix #11
On page refresh (or first visit), when there is no snapshot base, all fields will always be changed, but we want to keep the page parameter intact from the URL. When we **do** want to reset the page parameter (e.g. on a sorting change), we want to intercept the transition before `model` gets called, and replae the in progress transition with one where the page parameter is undefined, so it is also reflected in the URL. We can't do this in the model as that would make us do 2 search requests.