Downgrade Ember Data to v4.11 (#43)
v4.12.0 includes a relationship promise caching bug which causes the wrong data to be loaded. This has the side affect that we try to load the submission form with the wrong id which causes the app to break.

Downgrading to v4.11 fixes the issue for now. Once the bug is fixed we can update to 4.12 again.